Split Element type to it's own definition file (#8206)
